requirements.txt
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

boto3==1.34.162

google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1

pypdf==4.3.1
reportlab==4.2.2
pdfplumber==0.11.4

fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4

sqlalchemy==2.0.32

psycopg2-binary==2.9.9
google-auth-httplib2==0.2.0

google-cloud-secret-manager==2.26.0